requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
gunicorn
whitenoise
Pillow
dj-database-url
psycopg2-binary
